Revenue RevenueMonYearDTO Type 변경

+++ src/main/java/com/receipt/www/receiptbackend/revenue/command/application/dto/RevenueMonYearDTO.java
@@ -2,19 +2,19 @@ package com.receipt.www.receiptbackend.revenue.command.application.dto;
 
 public class RevenueMonYearDTO {
     private String monYearDate;
-    private int monYearDay;
-    private int monYearTotal;
-    private int monYearSale;
-    private int monYearActual;
-    private int monYearReceiptNum;
-    private int monYearReceiptPrice;
-    private int monYearReceiptValue;
-    private int monYearlySurtax;
+    private Long monYearDay;
+    private Long monYearTotal;
+    private Long monYearSale;
+    private Long monYearActual;
+    private Long monYearReceiptNum;
+    private Long monYearReceiptPrice;
+    private Long monYearReceiptValue;
+    private Long monYearlySurtax;
 
     public RevenueMonYearDTO() {
     }
 
-    public RevenueMonYearDTO(String monYearDate, int monYearDay, int monYearTotal, int monYearSale, int monYearActual, int monYearReceiptNum, int monYearReceiptPrice, int monYearReceiptValue, int monYearlySurtax) {
+    public RevenueMonYearDTO(String monYearDate, Long monYearDay, Long monYearTotal, Long monYearSale, Long monYearActual, Long monYearReceiptNum, Long monYearReceiptPrice, Long monYearReceiptValue, Long monYearlySurtax) {
         this.monYearDate = monYearDate;
         this.monYearDay = monYearDay;
         this.monYearTotal = monYearTotal;
@@ -34,67 +34,67 @@ public class RevenueMonYearDTO {
         this.monYearDate = monYearDate;
     }
 
-    public int getMonYearDay() {
+    public Long getMonYearDay() {
         return monYearDay;
     }
 
-    public void setMonYearDay(int monYearDay) {
+    public void setMonYearDay(Long monYearDay) {
         this.monYearDay = monYearDay;
     }
 
-    public int getMonYearTotal() {
+    public Long getMonYearTotal() {
         return monYearTotal;
     }
 
-    public void setMonYearTotal(int monYearTotal) {
+    public void setMonYearTotal(Long monYearTotal) {
         this.monYearTotal = monYearTotal;
     }
 
-    public int getMonYearSale() {
+    public Long getMonYearSale() {
         return monYearSale;
     }
 
-    public void setMonYearSale(int monYearSale) {
+    public void setMonYearSale(Long monYearSale) {
         this.monYearSale = monYearSale;
     }
 
-    public int getMonYearActual() {
+    public Long getMonYearActual() {
         return monYearActual;
     }
 
-    public void setMonYearActual(int monYearActual) {
+    public void setMonYearActual(Long monYearActual) {
         this.monYearActual = monYearActual;
     }
 
-    public int getMonYearReceiptNum() {
+    public Long getMonYearReceiptNum() {
         return monYearReceiptNum;
     }
 
-    public void setMonYearReceiptNum(int monYearReceiptNum) {
+    public void setMonYearReceiptNum(Long monYearReceiptNum) {
         this.monYearReceiptNum = monYearReceiptNum;
     }
 
-    public int getMonYearReceiptPrice() {
+    public Long getMonYearReceiptPrice() {
         return monYearReceiptPrice;
     }
 
-    public void setMonYearReceiptPrice(int monYearReceiptPrice) {
+    public void setMonYearReceiptPrice(Long monYearReceiptPrice) {
         this.monYearReceiptPrice = monYearReceiptPrice;
     }
 
-    public int getMonYearReceiptValue() {
+    public Long getMonYearReceiptValue() {
         return monYearReceiptValue;
     }
 
-    public void setMonYearReceiptValue(int monYearReceiptValue) {
+    public void setMonYearReceiptValue(Long monYearReceiptValue) {
         this.monYearReceiptValue = monYearReceiptValue;
     }
 
-    public int getMonYearlySurtax() {
+    public Long getMonYearlySurtax() {
         return monYearlySurtax;
     }
 
-    public void setMonYearlySurtax(int monYearlySurtax) {
+    public void setMonYearlySurtax(Long monYearlySurtax) {
         this.monYearlySurtax = monYearlySurtax;
     }
